+++ CHANGES.md
@@ -18,11 +18,11 @@ of that change.
 ### Changed:
 
 - Language models now use flattened loss, instead of flattening y in data loader
+- `ItemList.from_folder` now has an `include` parameter to only include certain folders
 
 ### Fixed:
 
 - `Learner.load` won't throw an error when trying to load an optimizer state of the wrong size, and silently ignore that optimizer state loading
-
 
 
 ## 1.0.32 (2018-12-02)